Move `compressor.cpp|h` and `envelope.cpp/h` to `src/audio`
