+++ src/audio/envelope.cpp
