+++ src/audio/envelope.h
