+++ src/audio/meson.build
@@ -3,6 +3,9 @@ libaudio_sources = files(
     'clap/library.cpp',
     'clap/plugin.cpp',
     'clap/plugin_manager.cpp',
+
+    'compressor.cpp',
+    'envelope.cpp',
 )
 
 libaudio = static_library(
@@ -10,6 +13,7 @@ libaudio = static_library(
     libaudio_sources,
     include_directories: incdir,
     dependencies: [
+        libiir_dep,
         libloguru_dep,
     ],
     cpp_args: warnings,
