+++ src/hardware/CMakeLists.txt
@@ -24,10 +24,8 @@ add_library(libhardware STATIC
 		adlib_gold.cpp
 		cmos.cpp
 		covox.cpp
-		compressor.cpp
 		disney.cpp
 		dma.cpp
-		envelope.cpp
 		gameblaster.cpp
 		gus.cpp
 		ide.cpp
@@ -86,6 +84,7 @@ find_package(SDL2_net REQUIRED)
 target_link_libraries(libhardware PRIVATE $<IF:$<TARGET_EXISTS:SDL2_net::SDL2_net>,SDL2_net::SDL2_net,SDL2_net::SDL2_net-static>)
 
 target_link_libraries(libhardware PRIVATE
+		libaudio
 		libesfmu
 		libresidfp
 		libym7128bemu
