+++ src/hardware/meson.build
@@ -24,10 +24,8 @@ libhardware_sources = files(
     'adlib_gold.cpp',
     'cmos.cpp',
     'covox.cpp',
-    'compressor.cpp',
     'disney.cpp',
     'dma.cpp',
-    'envelope.cpp',
     'gameblaster.cpp',
     'gus.cpp',
     'ide.cpp',
