+++ vs/dosbox.vcxproj
@@ -596,6 +596,8 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\audio\clap\library.cpp" />
     <ClCompile Include="..\src\audio\clap\plugin.cpp" />
     <ClCompile Include="..\src\audio\clap\plugin_manager.cpp" />
+    <ClCompile Include="..\src\audio\compressor.cpp" />
+    <ClCompile Include="..\src\audio\envelope.cpp" />
     <ClCompile Include="..\src\dosbox.cpp" />
     <ClCompile Include="..\src\dos\cdrom_win32.cpp" />
     <ClCompile Include="..\src\libs\PDCurses\pdcurses\addch.c" />
@@ -732,11 +734,9 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\gui\titlebar.cpp" />
     <ClCompile Include="..\src\hardware\adlib_gold.cpp" />
     <ClCompile Include="..\src\hardware\cmos.cpp" />
-    <ClCompile Include="..\src\hardware\compressor.cpp" />
     <ClCompile Include="..\src\hardware\covox.cpp" />
     <ClCompile Include="..\src\hardware\disney.cpp" />
     <ClCompile Include="..\src\hardware\dma.cpp" />
-    <ClCompile Include="..\src\hardware\envelope.cpp" />
     <ClCompile Include="..\src\hardware\gameblaster.cpp" />
     <ClCompile Include="..\src\hardware\gus.cpp" />
     <ClCompile Include="..\src\hardware\ide.cpp" />
@@ -924,7 +924,6 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\include\dos_system.h" />
     <ClInclude Include="..\include\dosbox.h" />
     <ClInclude Include="..\include\drives.h" />
-    <ClInclude Include="..\include\envelope.h" />
     <ClInclude Include="..\include\ethernet.h" />
     <ClInclude Include="..\include\fpu.h" />
     <ClInclude Include="..\include\fraction.h" />
@@ -986,6 +985,8 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\src\audio\clap\library.h" />
     <ClInclude Include="..\src\audio\clap\plugin.h" />
     <ClInclude Include="..\src\audio\clap\plugin_manager.h" />
+    <ClInclude Include="..\src\audio\compressor.h" />
+	<ClInclude Include="..\src\audio\envelope.h" />
     <ClInclude Include="..\src\capture\capture.h" />
     <ClInclude Include="..\src\capture\capture_audio.h" />
     <ClInclude Include="..\src\capture\capture_midi.h" />
@@ -1087,7 +1088,6 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\src\gui\shader_manager.h" />
     <ClInclude Include="..\src\gui\titlebar.h" />
     <ClInclude Include="..\src\hardware\adlib_gold.h" />
-    <ClInclude Include="..\src\hardware\compressor.h" />
     <ClInclude Include="..\src\hardware\covox.h" />
     <ClInclude Include="..\src\hardware\disney.h" />
     <ClInclude Include="..\src\hardware\gameblaster.h" />
